requirements.txt
Flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
passlib[argon2]

vosk
cmudict

torch
transformers
librosa
soundfile
numpy
pydub
gTTS

python-dotenv
google-genai